requirements.txt
requests>=2.31.0
pycryptodome>=3.21.0,<4.0.0
tenacity>=8.2.3,<9.0.0
typing-extensions>=4.5.0,<5.0.0
